[Sign Up][Feature] Change upload photo to take photo
